# [SPARK-31705][SQL] Push more possible predicates through Join via CNF conversion