--- a/sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/expressions/predicates.scala
+++ b/sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/expressions/predicates.scala
@@ -241,8 +241,6 @@ trait PredicateHelper {
       }
       resultStack.push(cnf)
     }
-    assert(resultStack.length == 1,
-      s"Fail to convert expression ${condition} to conjunctive normal form")
     resultStack.top
   }