@gengliangwang 2020-06-05
As I talked to @wangyum offline, I am taking https://github.com/apache/spark/pull/28575 over for the CNF implementation and config naming. 

There have been PRs for CNF conversion, such as https://github.com/apache/spark/pull/10444, https://github.com/apache/spark/pull/15558, #28575. The common issue is the recursive implementation can slow, or even cause a stack overflow exception.

With this non-recursive implementation, the rule should be faster and more robust.