@gengliangwang 2020-06-06
test failure from TPCDSQuerySuite:
> 18884 was not less than or equal to 8000 too long generated codes found in the WholeStageCodegenExec subtree (id=375762) and JIT optimization might not work:

I will update the PR to simplify the pushed down predicates