@gengliangwang 2020-06-06
I find a simpler solution and create https://github.com/apache/spark/pull/28741.
Close this one now.