@maropu 2020-06-11
(Just to check) Btw, it seems the previous works about CNF tried to implement this conversion in an independent rule for `Filter` plans  (e.g., the @viirya one: https://github.com/apache/spark/pull/15558/files#diff-a1acb054bc8888376603ef510e6d0ee0R139). On the other hand, this PR only targets at join queries. Is this because this conversion has a severe trade-off relationship between time complexity / performance gains, but join quries can get much performance improvements even in this case? Is my understanding correct?