@gengliangwang 2020-06-11
@maropu Yes pushing down predicates through join should be the major scenario.