@AngersZhuuuu 2020-06-11
@maropu 
After this merge I will raise a pr base on his non-recursive code  to improve data scan's partition pruning.
https://github.com/apache/spark/pull/24973